Almost finished role-based login. Have to do some polish and implement some checks in the main layout and IsAdmin
