+++ AutomobileRepairShop/Program.cs
@@ -10,7 +10,7 @@ builder.Services.AddControllersWithViews();
 builder.Services.AddAuthentication(CookieAuthenticationDefaults.AuthenticationScheme)
     .AddCookie(options =>
     {
-        options.ExpireTimeSpan = TimeSpan.FromMinutes(0);
+        options.ExpireTimeSpan = TimeSpan.FromMinutes(1);
         options.LoginPath = "/Home";
         options.Cookie.Name = "userSession";
         //options.SlidingExpiration = true;
