+++ AutomobileRepairShop/Views/Account/Index.cshtml
@@ -1,5 +1,9 @@
 ﻿@*
     For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 *@
-@{
-}
+
+
+<div class="text-center">
+    <h1 class="display-4">HELLU USER</h1>
+    @ViewBag.
+</div>
