+++ AutomobileRepairShop/Views/Shared/Login.cshtml
@@ -4,6 +4,7 @@
 @model AutomobileRepairShop.Models.User
 
 @{
+ViewBag.Title = "Login";
 <!DOCTYPE html>
 <html>
 <section class="vh-100">
@@ -18,7 +19,7 @@
 
                 <div class="d-flex align-items-center h-custom-2 px-5 ms-xl-4 mt-5 pt-5 pt-xl-0 mt-xl-n5">
 
-                    <form asp-controller="Login" asp-action="Login" method="post" style="width: 23rem;">
+                    <form asp-controller="Home" asp-action="Login" method="post" style="width: 23rem;">
 
                         <h3 class="fw-normal mb-3 pb-3" style="letter-spacing: 1px;">Log in</h3>
