+++ AutomobileRepairShop/Views/Shared/Register.cshtml
@@ -19,7 +19,7 @@
                 <div class="d-flex align-items-center h-custom-2 px-5 ms-xl-4 mt-5 pt-5 pt-xl-0 mt-xl-n5">
                     
                     <!-- registration form -->
-                    <form asp-controller="Registration" asp-action="Register" method="post" style="width: 23rem;">
+                    <form asp-controller="Home" asp-action="Register" method="post" style="width: 23rem;">
 
                         <!-- "Register" title -->
                         <h3 class="fw-normal mb-3 pb-3" style="letter-spacing: 1px;">Register</h3>
