+++ AutomobileRepairShop/Views/Shared/_Layout.cshtml
@@ -1,4 +1,5 @@
-﻿<!DOCTYPE html>
+﻿
+<!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="utf-8" />
@@ -7,7 +8,7 @@
     <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" />
     <link rel="stylesheet" href="~/css/site.css" asp-append-version="true" />
     <link rel="stylesheet" href="~/AutomobileRepairShop.styles.css" asp-append-version="true" />
-
+    
 </head>
 <body>
     <header>
@@ -25,6 +26,8 @@
                         </li>
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
+                        
+                        <!--Admin dashboard-->  
                         <li class="nav-item dropdown">
                             <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
                                 Admin dashboard
@@ -36,12 +39,16 @@
                         </li>
                     </ul>
                     <ul class="navbar-nav ms-auto flex-nowrap">
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Login">Login</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Register">Register</a>
-                        </li>
+
+                        <!--Login, Register, Logout-->
+                        @if (@ViewBag.IsLogged == false) //Login Register
+                        {
+                            <li class="nav-item"><a id="login" class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Login">Login</a></li>
+                            <li class="nav-item"><a id="register" class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Register">Register</a></li>
+                        }
+                        else @if(@ViewBag.IsLogged == true){ //Logout
+                        <li class="nav-item"><a id="logout" class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Login">Logout</a></li>
+                        }
                     </ul>
 
                 </div>
